Reorganize and document queryset module
